+++ CHANGELOG.md
@@ -5,3 +5,15 @@
 ### First initialization
 
 * First initialization
+
+# 1.0.2 (2023-07-19)
+
+### Added
+
+* getRelease() - Getting the release version
+* getSession() - Getting a session
+* getQRCode(clientId) - Getting wg-easy client qr-code
+* updateAddress(clientId, address) - Update wg-easy client address
+
+### Fix
+* rename(clientId, newName) - now it is not possible to take an existing name